refactor: split JetStream bus into focused modules (#1221)
